Move interface helper functions to `qml/math/interface_utils.py` (#6649)
**Context:**

Interface handling is currently scattered throughout the codebase. To
standardize supported interfaces and streamline workflows, the first
step is to centralize all related functions. `qml.math` is the natural
choice as most of the interface management is currently being done
through helpers found in `qml/math/utils.py`.

**Description of the Change:**

All interface handling functions independent of Pennylane structures
(tapes, qnodes, etc) have been moved to `qml/math/interface_utils.py`.

Added new unit test file `test_resolve_interface.py` to improve testing
of interface conversion logic in `execution.py`.

**Benefits:**

- Improved code organization.
- Easier to maintain and update interface handling logic.

**Possible Drawbacks:**

None identified.

[sc-79357]

---------
