+++ pennylane/math/__init__.py
@@ -89,11 +89,16 @@ from .utils import (
     cast,
     cast_like,
     convert_like,
+    in_backprop,
+    requires_grad,
+    is_abstract,
+)
+from .interface_utils import (
+    get_canonical_interface_name,
+    SupportedInterfaceUserInput,
+    SUPPORTED_INTERFACE_NAMES,
     get_deep_interface,
     get_interface,
-    in_backprop,
-    is_abstract,
-    requires_grad,
 )
 
 sum = ar.numpy.sum
@@ -159,6 +164,7 @@ __all__ = [
     "frobenius_inner_product",
     "get_dtype_name",
     "get_interface",
+    "get_canonical_interface_name",
     "get_deep_interface",
     "get_trainable_indices",
     "in_backprop",
