+++ pennylane/math/multi_dispatch.py
@@ -24,7 +24,8 @@ from autoray import numpy as np
 from numpy import ndarray
 
 from . import single_dispatch  # pylint:disable=unused-import
-from .utils import cast, cast_like, get_interface, requires_grad
+from .interface_utils import get_interface
+from .utils import cast, cast_like, requires_grad
 
 
 # pylint:disable=redefined-outer-name
