+++ tests/qnn/test_qnn_torch.py
@@ -644,7 +644,7 @@ def test_qnode_interface_not_mutated(interface):
     assert (
         qlayer.qnode.interface
         == circuit.interface
-        == qml.workflow.execution.INTERFACE_MAP[interface]
+        == qml.math.get_canonical_interface_name(interface)
     )
